Extract internal/tokenfile package from deviceauth
Move Token type, Load, and Save functions into a dedicated
internal/tokenfile package to separate token file I/O from
the device authorization flow logic.
