+++ internal/command/task_create.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/icholy/xagent/internal/deviceauth"
 	xagentv1 "github.com/icholy/xagent/internal/proto/xagent/v1"
+	"github.com/icholy/xagent/internal/tokenfile"
 	"github.com/icholy/xagent/internal/xagentclient"
 	"github.com/urfave/cli/v3"
 )
@@ -54,7 +54,7 @@ var TaskCreateCommand = &cli.Command{
 	},
 	Action: func(ctx context.Context, cmd *cli.Command) error {
 		serverURL := cmd.String("server")
-		token, err := deviceauth.LoadToken(cmd.String("token-file"))
+		token, err := tokenfile.Load(cmd.String("token-file"))
 		if err != nil {
 			return fmt.Errorf("failed to load token: %w", err)
 		}
