+++ internal/deviceauth/deviceauth.go
@@ -3,9 +3,7 @@ package deviceauth
 import (
 	"cmp"
 	"context"
-	"encoding/json"
 	"fmt"
-	"os"
 	"time"
 
 	"github.com/zitadel/oidc/v3/pkg/client/rp"
@@ -13,42 +11,6 @@ import (
 )
 
 var scopes = []string{oidc.ScopeOpenID, oidc.ScopeProfile, oidc.ScopeEmail}
-
-// Token stores the API key
-type Token struct {
-	APIKey string `json:"api_key"`
-}
-
-// Valid reports whether the token has a non-empty API key
-func (t *Token) Valid() bool {
-	return t != nil && t.APIKey != ""
-}
-
-// LoadToken reads a token from a JSON file.
-// Returns a non-nil Token even if the file doesn't exist (with an empty API key).
-func LoadToken(path string) (*Token, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return &Token{}, nil
-		}
-		return nil, err
-	}
-	var token Token
-	if err := json.Unmarshal(data, &token); err != nil {
-		return nil, err
-	}
-	return &token, nil
-}
-
-// SaveToken writes a token to a JSON file.
-func SaveToken(path string, token *Token) error {
-	data, err := json.MarshalIndent(token, "", "  ")
-	if err != nil {
-		return err
-	}
-	return os.WriteFile(path, data, 0600)
-}
 
 // DeviceFlowOptions configures the device authorization flow.
 type DeviceFlowOptions struct {
